requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
SQLAlchemy==2.0.35
requests==2.31.0
PyYAML==6.0.1
pytest==7.4.3
pytest-cov==4.1.0